+++ dds/src/implementation/mod.rs
@@ -4,4 +4,4 @@ pub mod data_representation_builtin_endpoints;
 pub mod data_representation_inline_qos;
 pub mod payload_serializer_deserializer;
 pub mod rtps;
-pub mod udp;
+pub mod udp_transport;
