--- dds/src/implementation/udp/mod.rs
@@ -1,1 +1,0 @@
-pub mod udp_transport;
